optimize: change column and vcolumn to unsigned short
The crash project doesn't have lines longer than 64K characters, so
we should be fine with using the short type. This helps reduce the
memory requirements, because it shrinks node_t.

+++ parser.h
@@ -75,7 +75,7 @@ struct parsed_file;
 typedef struct {
 	struct parsed_file *pf;
 	int line;
-	int column, vcolumn;
+	unsigned short column, vcolumn;
 	struct dynstr *text;
 } pos_t;
